+++ src/naksu/mebroutines/mebroutines.go
@@ -7,10 +7,8 @@ import (
 	"io"
 	golog "log"
 	"os"
-	"os/exec"
 	"path/filepath"
 	"runtime"
-	"strings"
 
 	"naksu/log"
 	"naksu/xlate"
@@ -27,33 +25,6 @@ func Close(c io.Closer) {
 	if err != nil {
 		golog.Fatal(err)
 	}
-}
-
-// RunAndGetOutput runs command with arguments and returns output as a string
-func RunAndGetOutput(commandArgs []string, logAction bool) (string, error) {
-	if logAction {
-		log.Debug(fmt.Sprintf("RunAndGetOutput: %s", strings.Join(commandArgs, " ")))
-	}
-
-	/* #nosec */
-	cmd := exec.Command(commandArgs[0], commandArgs[1:]...)
-
-	out, err := cmd.CombinedOutput()
-
-	if err != nil {
-		log.Debug(fmt.Sprintf("command failed: %s (%v)", strings.Join(commandArgs, " "), err))
-	}
-
-	if out != nil {
-		if logAction {
-			log.Debug("RunAndGetOutput returns combined STDOUT and STDERR:")
-			log.Debug(string(out))
-		}
-	} else {
-		log.Debug("RunAndGetOutput returned NIL as combined STDOUT and STDERR")
-	}
-
-	return string(out), err
 }
 
 func getFileMode(path string) (os.FileMode, error) {
